% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_coord.R
\name{bin_coord}
\alias{bin_coord}
\title{Optional spatial binning}
\usage{
bin_coord(expr, coord, num_bin = 100)
}
\arguments{
\item{expr}{The normalized gene expression matrix. Rows represent genes and columns represent spots/cells}

\item{coord}{The matrix of spatial locations. Rows represent spots/cells. The column named "row" ("col") represents row (column) coordinates}

\item{num_bin}{Number of bins in each spatial dimension (100 by default)}
}
\value{
A list of binned \code{expr} and \code{coord}
}
\description{
Optional spatial binning
}
