% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_pseudotime.R
\name{bin_pseudotime}
\alias{bin_pseudotime}
\title{Optional temporal binning}
\usage{
bin_pseudotime(expr, pseudotime, num_bin = 100)
}
\arguments{
\item{expr}{The normalized gene expression matrix. Rows represent genes and columns represent cells}

\item{pseudotime}{The vector of user-provided pseudotime values}

\item{num_bin}{Number of bins (100 by default)}
}
\value{
A list of binned \code{expr} and \code{pseudotime}
}
\description{
Optional temporal binning
}
