% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreTSA_spatial.R
\name{spatialFit}
\alias{spatialFit}
\title{Fit the gene expression along spatial locations}
\usage{
spatialFit(expr, coord, knot = 0, maxknotallowed = 5)
}
\arguments{
\item{expr}{The normalized gene expression matrix. Rows represent genes and columns represent spots/cells}

\item{coord}{The matrix of spatial locations. Rows represent spots/cells. The column named "row" ("col") represents row (column) coordinates}

\item{knot}{Number of knots (0 by default) or \code{"auto"} for automatic selection}

\item{maxknotallowed}{A user-defined maximum number of knots (5 by default)}
}
\value{
The fitted expression matrix. Rows represent genes and columns represent spots/cells
}
\description{
Fit the gene expression along spatial locations
}
