% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreTSA_spatial.R
\name{spatialTest_bin}
\alias{spatialTest_bin}
\title{Test if the gene expression is associated with spatial locations (binning)}
\usage{
spatialTest_bin(expr, coord, knot = 0, maxknotallowed = 5, numbin = 10)
}
\arguments{
\item{expr}{The normalized gene expression matrix. Rows represent genes and columns represent spots/cells}

\item{coord}{The matrix of spatial locations. Rows represent spots/cells. The column named "row" ("col") represents row (column) coordinates}

\item{knot}{Number of knots (0 by default) or \code{"auto"} for automatic selection}

\item{maxknotallowed}{A user-defined maximum number of knots (5 by default)}

\item{numbin}{Number of bins (10 by default)}
}
\value{
A data frame with the p-value (FDR) and test statistic for each gene (each row)
}
\description{
Test if the gene expression is associated with spatial locations (binning)
}
