% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreTSA_temporal.R
\name{temporalFit_bin}
\alias{temporalFit_bin}
\title{Fit the gene expression along pseudotime values (binning)}
\usage{
temporalFit_bin(expr, pseudotime, knot = 0, maxknotallowed = 10, numbin = 10)
}
\arguments{
\item{expr}{The normalized gene expression matrix. Rows represent genes and columns represent cells}

\item{pseudotime}{The vector of user-provided pseudotime values}

\item{knot}{Number of knots (0 by default) or \code{"auto"} for automatic selection}

\item{maxknotallowed}{A user-defined maximum number of knots (10 by default)}

\item{numbin}{Number of bins (10 by default)}
}
\value{
The fitted expression matrix. Rows represent genes and columns represent cells
}
\description{
Fit the gene expression along pseudotime values (binning)
}
