% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreTSA_temporal.R
\name{temporalTest_bin}
\alias{temporalTest_bin}
\title{Test if the gene expression is associated with pseudotime values (binning)}
\usage{
temporalTest_bin(
  expr,
  pseudotime,
  pseudotime_permute = NULL,
  knot = 0,
  maxknotallowed = 10,
  numbin = 10
)
}
\arguments{
\item{expr}{The normalized gene expression matrix. Rows represent genes and columns represent cells}

\item{pseudotime}{The vector of user-provided pseudotime values}

\item{pseudotime_permute}{A list of permuted pseudotime values from subsampled cells. Each element in the list has the same format of the argument \code{pseudotime}}

\item{knot}{Number of knots (0 by default) or \code{"auto"} for automatic selection}

\item{maxknotallowed}{A user-defined maximum number of knots (10 by default)}

\item{numbin}{Number of bins (10 by default)}
}
\value{
A data frame with the p-value (FDR) and test statistic for each gene (each row)
}
\description{
Test if the gene expression is associated with pseudotime values (binning)
}
